@@ -2,28 +2,27 @@
   "name": "cssgridgen",
   "version": "0.1.0",
   "private": true,
+  "type": "module",
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview",
     "test": "jest"
   },
   "dependencies": {
     "core-js": "^2.6.5",
     "focus-trap": "^5.0.1",
-    "vue": "^2.6.10",
-    "vue-i18n": "^8.11.2",
-    "vuex": "^3.0.1"
+    "vue": "^3.5.20",
+    "vue-i18n": "^11.1.11",
+    "vuex": "^4.0.2"
   },
   "devDependencies": {
-    "@babel/core": "^7.6.2",
-    "@babel/preset-env": "^7.6.2",
-    "@vue/cli-plugin-babel": "^3.8.0",
-    "@vue/cli-service": "^3.8.0",
-    "babel-jest": "^24.9.0",
+    "@vitejs/plugin-vue": "^6.0.1",
+    "autoprefixer": "^10.4.21",
     "jest": "^24.9.0",
-    "sass": "^1.18.0",
-    "sass-loader": "^7.1.0",
-    "vue-template-compiler": "^2.6.10"
+    "postcss": "^8.5.6",
+    "sass-embedded": "^1.91.0",
+    "vite": "^7.1.3"
   },
   "postcss": {
     "plugins": {
