@@ -6,14 +6,14 @@
     <p>
       {{ $t('explain.paragraph4') }}
       <a
-        href="https://twitter.com/rachelandrew"
-        target="_blank"
+          href="https://twitter.com/rachelandrew"
+          target="_blank"
       >Rachel Andrew</a>,
       <a href="https://twitter.com/jensimmons" target="_blank">Jen Simmons</a>, {{ $t('utils.and') }}
       <a href="https://twitter.com/geddski" target="_blank">Dave Geddes</a> {{ $t('explain.paragraph5') }}
       <a
-        href="https://css-tricks.com/snippets/css/complete-guide-grid/"
-        target="_blank"
+          href="https://css-tricks.com/snippets/css/complete-guide-grid/"
+          target="_blank"
       >{{ $t('explain.cssguide') }}</a>, {{ $t('explain.paragraph6') }}
       <a href="https://cssgridgarden.com/" target="_blank">Grid Garden</a> {{ $t('explain.paragraph7') }}
     </p>
@@ -21,8 +21,8 @@
     <p class="cursive">
       {{ $t('explain.contributions') }}
       <a
-        href="https://github.com/sdras/cssgridgenerator"
-        target="_blank"
+          href="https://github.com/sdras/cssgridgenerator"
+          target="_blank"
       >{{ $t('explain.fork') }}</a>
     </p>
   </section>
@@ -33,11 +33,14 @@ export default {};
 </script>
 
 <style lang="scss" scoped>
+@use "../styles/variables.scss" as root;
+
 .cursive {
   font-family: "Caveat", cursive;
   font-size: 25px;
   color: #ffd908;
   font-weight: normal;
+
   a,
   a:visited,
   a:active {
@@ -53,7 +56,7 @@ strong {
 a,
 a:visited,
 a:active {
-  color: $teal;
+  color: root.$teal;
   text-decoration: none;
 }
 </style>
